+++ binanceenv/spaces.py
@@ -40,7 +40,7 @@ def get_action_space_obj(action_type='discrete'):
 
 class IndicatorsSpace:
     def __init__(self, ind_num):
-        self.__observation_space = spaces.Box(low=0.0, high=1.0, shape=(ind_num,), dtype=np.float32, seed=42)
+        self.__observation_space = spaces.Box(low=0.0, high=1.0, shape=(ind_num,), dtype=np.float64, seed=42)
         self.name = 'indicators'
 
     @property
@@ -58,9 +58,9 @@ class IndicatorsAndAssetsSpace:
         # high = np.ones((assets_num + ind_num,))
         # low[:assets_num] = 0.0
         # high[:assets_num] = 1.0
-        self.__observation_space = spaces.Box(low=0.0, high=1.0, shape=(ind_num + assets_num,), dtype=np.float32,
+        self.__observation_space = spaces.Box(low=0.0, high=1.0, shape=(ind_num + assets_num,), dtype=np.float64,
                                               seed=42)
-        # self.__observation_space = Box(low=low, high=high, dtype=np.float32, seed=42)
+        # self.__observation_space = Box(low=low, high=high, dtype=np.float64, seed=42)
         self.name = 'indicators_assets'
 
     @property
@@ -77,7 +77,7 @@ class AssetsCloseIndicatorsSpace:
         self.__observation_space = spaces.Box(low=low,
                                               high=high,
                                               shape=(ind_num + assets_num + 1,),
-                                              dtype=np.float32,
+                                              dtype=np.float64,
                                               seed=42)
         self.name = 'assets_close_indicators'
 
@@ -95,7 +95,7 @@ class AssetsCloseActionIndicatorsSpace:
         self.__observation_space = spaces.Box(low=low,
                                               high=high,
                                               shape=(assets_data + actions + ind_num, ),
-                                              dtype=np.float32,
+                                              dtype=np.float64,
                                               seed=42)
         self.name = 'assets_close_action_indicators'
 
@@ -113,7 +113,7 @@ class LookbackAssetsCloseIndicatorsSpace:
         self.__observation_space = spaces.Box(low=low,
                                               high=high,
                                               shape=((ind_num + assets_data + 1) * lookback,),
-                                              dtype=np.float32,
+                                              dtype=np.float64,
                                               seed=42)
         self.name = 'lookback_assets_close_indicators'
 
@@ -131,7 +131,7 @@ class LookbackAssetsCloseIndicatorsSpaceCNN:
         self.__observation_space = spaces.Box(low=low,
                                               high=high,
                                               shape=(lookback, assets_data + ind_num),
-                                              dtype=np.float32,
+                                              dtype=np.float64,
                                               seed=42)
         self.name = 'lookback_assets_close_indicators_cnn'
 
@@ -149,7 +149,7 @@ class LookbackAssetsCloseActionIndicatorsSpaceCNN:
         self.__observation_space = spaces.Box(low=low,
                                               high=high,
                                               shape=(lookback, assets_data + actions + ind_num),
-                                              dtype=np.float32,
+                                              dtype=np.float64,
                                               seed=42)
         self.name = 'lookback_assets_close_action_indicators'
 
@@ -167,7 +167,7 @@ class LookbackAssetsCloseIndicatorsActionSpace:
         self.__observation_space = spaces.Box(low=low,
                                               high=high,
                                               shape=(lookback, ind_num + assets_data + 1 + actions),
-                                              dtype=np.float32,
+                                              dtype=np.float64,
                                               seed=42)
         self.name = 'lookback_assets_close_indicators_action'
 
@@ -184,11 +184,11 @@ class LookbackDictOHLCAssetsIndicatorsSpace:
     def __init__(self, ind_num, assets_num, lookback):
         self.__observation_space = spaces.Dict(
             {"assets": spaces.Box(low=0.0, high=1.0, shape=(5 * assets_num,),
-                                  dtype=np.float32, seed=42),
-             "ohlc": spaces.Box(low=0.0, high=1.0, shape=(4, lookback), dtype=np.float32,
+                                  dtype=np.float64, seed=42),
+             "ohlc": spaces.Box(low=0.0, high=1.0, shape=(4, lookback), dtype=np.float64,
                                 seed=42),
              "indicators": spaces.Box(low=0.0, high=1.0, shape=(ind_num, lookback),
-                                      dtype=np.float32, seed=42),
+                                      dtype=np.float64, seed=42),
              })
         self.name = 'lookback_dict'
 
@@ -207,7 +207,7 @@ class IndicatorsAndPNLSpace:
         high = np.ones((pnl_num + ind_num,))
         low[:pnl_num] = -1.0
         high[:pnl_num] = 3.0
-        self.__observation_space = spaces.Box(low=low, high=high, dtype=np.float32, seed=42)
+        self.__observation_space = spaces.Box(low=low, high=high, dtype=np.float64, seed=42)
         self.name = 'indicators_pnl'
 
     @property
@@ -267,7 +267,7 @@ class DiscreteActionSpaceSpot(DiscreteActionSpace):
 class BoxActionSpace:
     def __init__(self, n_action):
         self.n_action = n_action
-        self.__action_space = spaces.Box(low=0, high=1, shape=(n_action,), dtype=np.float32)
+        self.__action_space = spaces.Box(low=0, high=1, shape=(n_action,), dtype=np.float64)
         self.name = 'box'
 
     def convert2action(self, action, masked_actions=None) -> Tuple[float, float]:
@@ -289,7 +289,7 @@ class BoxActionSpace:
 
 class BoxExtActionSpace:
     def __init__(self, n_action):
-        self.__action_space = spaces.Box(low=-1., high=1., shape=(n_action,), dtype=np.float32)
+        self.__action_space = spaces.Box(low=-1., high=1., shape=(n_action,), dtype=np.float64)
         self.name = 'box1_1'
 
     @staticmethod
@@ -319,7 +319,7 @@ class ActionsBins:
         self.box_range = (np.max(box) - np.min(box))
         self.n_actions = n_actions
         self.step = self.box_range / n_actions
-        self.bins = np.arange(box[0], box[1] + 1e-7, step=self.step, dtype=np.float32)
+        self.bins = np.arange(box[0], box[1] + 1e-7, step=self.step, dtype=np.float64)
         self.pairs = [(self.bins[ix - 1], self.bins[ix]) for ix in range(1, len(self.bins))]
 
     def bins_2actions(self, value) -> Tuple[int, float]:
@@ -342,8 +342,8 @@ class ActionsBins:
 
 class SellBuyHoldAmount:
     def __init__(self):
-        self.__action_space = spaces.Box(low=np.array([-1.0, -1.0]), high=np.array([1.0, 1.0]), dtype=np.float32)
-        # self.__action_space = Box(low=-1., high=1., shape=(2,), dtype=np.float32)
+        self.__action_space = spaces.Box(low=np.array([-1.0, -1.0]), high=np.array([1.0, 1.0]), dtype=np.float64)
+        # self.__action_space = Box(low=-1., high=1., shape=(2,), dtype=np.float64)
         self.name = 'two_actions'
 
     def convert2action(self, action: np.ndarray, masked_actions=None):
@@ -373,7 +373,7 @@ class SellBuyHoldAmount:
 
 class BinBoxActionSpace:
     def __init__(self, n_action, low=-1., high=1.):
-        self.__action_space = spaces.Box(low=low, high=high, shape=(1,), dtype=np.float32)
+        self.__action_space = spaces.Box(low=low, high=high, shape=(1,), dtype=np.float64)
         self.actions_bins_obj = ActionsBins([low, high], n_action)
         self.name = 'binbox'
